jvm-attach: crate to implement JVM socket control protocol
JVMs implement a simple protocol to facilitate issuing commands
over a local socket. This is how the `jcmd` tool works.

This commit implements this client protocol for Hotspot derived
JVMs. Only UNIX support so far. But Windows should be pretty
easy to implement.

We probably still want to improve the API. But this code seems to
work and I'm able to issue commands to locally running JVMs. So
that's a start.