jfr-reader: rework lifetimes
We were using `'v: 'a` in a bunch of places, which was flat out wrong.
The returned Value should inherit the lifetime of the resolver, which
is implicitly shorter than `'a` since the resolver holds a reference
to `'a` data.

We purge the extra `'v` lifetime and use the resolver's own lifetime.