jfr-reader: rework lifetimes for EventRecord
We can't use a Cow for lifetime reasons.

We store a raw slice instead and change the trait to reflect the
original lifetime of slices. This way resolve_value() doesn't emit
a value with a lifetime dependence on self. It took me way too long
to debug this. I did not realize this behavior of Cow until today.
But it totally makes sense.