jfr-reader: handle constant index of 0 as null
Previously we were choking on constants expansions for various
references to constants at index 0. Turns out 0 is a special value
indicating a null value.

We change our Value enumeration to reflect this.

Since any Java value can be null, this has broader implications
for how events need to be modeled. But we'll cross that bridge
later.