jfr-reader: implement deserialization for Value and associated types
We want to support converting our generic Value into more specific,
typed structs. To do this we're going to leverage serde and its
generic deserialization support.

This commit implements deserializers for Value and associated types
(Object and Vec<Value>). There are still some unimplemented features
(such as string constant pool references and handling the common event
fields in a user-friendly way). But the foundation is there.

I iterated on this code a while before committing and anticipate we'll
be refactoring it in commits ahead. But for right now I believe
the deserialization is somewhat functional.