jfr-reader: define and implement an EventType trait
Now that we have structured event structs we want to start making them
more useful.

In order to map the JFR event to a Rust struct we'll need to find the
Rust struct from the JFR class name. To prepare for this, we define an
EventType trait holding the name of the event. We teach our metadata.xml
emitter to implement it.