jfr-reader: start to build out event time awareness
Event times are expressed in ticks. Those ticks require metadata
in the chunk header to resolve into wall time or time units.

This commit introduces functionality for performing time resolution.

We implement a new TimeResolver struct that holds metadata from the
chunk header and metadata and derives time values from ticks.

We also build out APIs on the chunk-level APIs to more easily access
start times. We even build out a convenience method for sorting events
in a chunk by their start time.

There's still work to be done to make retrieval of times simpler so
we don't need to pass in a TimeResolver. But we can figure this out
later.