jfr-reader: add benchmarks for JFR reading
We add a JFR recording from IntelliJ to provide an input corpus. We
then define various benchmarks for performing common actions on it.
The goal with this benchmark is to help us better track performance
of core functionality over time.

The benchmark results clearly indicate that typed deserialization is
slow. (But I knew that.)