jfr-reader: refactor string deserialization
We rework the ConstantResolver trait to be more generic for
mapping returned values. This lays the foundation for caching
although we stop short of implementing caching here.

The old string resolution mechanism has been ripped out and now
uses the new trait methods. As part of this, the deserializer
code for string constant pool references uses the new string
resolution API and dispatches directly to a string visitor without
going through a ValueDeserializer. Somewhat surprisingly this didn't
appear to have a meaningful impact on performance like I thought it
would!