jfr-reader: rework lifetimes of values
After renaming lifetimes from 'a to 'chunk, I immediately found some
areas where lifetimes were logically incorrect.

To correct this, I needed to introduce 'resolver and 'cr lifetimes to
refer to the event resolver and constants resolver instances. I believe
the new lifetimes are semantically more correct.

However, the lifetimes are still not great. For example, Values capture
the 'resolver lifetime because Object holds a reference to ClassElement.
This seems a bit questionable. I may revist this later. But at least
we're not mixing lifetimes anymore.