jfr-reader: store ClassElement as Arc<T>
Value currently have a lifetime on the resolver because of an internal
reference to the shared ClassElement. This is making lifetimes really
hard to reason about and makes callers have to contort themselves to
adhere to lifetimes.

As the first step towards eliminating this lifetime, we have the
resolver store ClassElement as an Arc<T>. This will enable us to return
a cheaply cloned owned value instead.